remove junk function in db models
