+++ packages/insomnia/src/ui/components/editors/body/body-editor.tsx
@@ -16,9 +16,6 @@ import type {
   Request,
   RequestBodyParameter,
 } from '../../../../models/request';
-import {
-  newBodyRaw,
-} from '../../../../models/request';
 import type { Workspace } from '../../../../models/workspace';
 import { NunjucksEnabledProvider } from '../../../context/nunjucks/nunjucks-enabled-context';
 import { AskModal } from '../../modals/ask-modal';
@@ -43,11 +40,25 @@ export const BodyEditor: FC<Props> = ({
   environmentId,
 }) => {
   const handleRawChange = useCallback((rawValue: string) => {
-    models.request.update(request, { body: newBodyRaw(rawValue, request.body.mimeType || '') });
+    models.request.update(request, {
+      body: typeof request.body.mimeType !== 'string' ? {
+        text: rawValue,
+      } : {
+        mimeType: request.body.mimeType.split(';')[0],
+        text: rawValue,
+      },
+    });
   }, [request]);
 
   const handleGraphQLChange = useCallback((content: string) => {
-    models.request.update(request, { body: newBodyRaw(content, CONTENT_TYPE_GRAPHQL) });
+    models.request.update(request, {
+      body: typeof CONTENT_TYPE_GRAPHQL !== 'string' ? {
+        text: content,
+      } : {
+        mimeType: CONTENT_TYPE_GRAPHQL.split(';')[0],
+        text: content,
+      },
+    });
   }, [request]);
 
   const handleFormUrlEncodedChange = useCallback((params: RequestBodyParameter[]) => {
